requirements.txt
streamlit==1.32.0
torch==2.1.2
diffusers==0.25.0
transformers==4.36.2
accelerate==0.26.1
pillow
speechrecognition
numpy
